Added protest dashboard navigator
