+++ src/screens/ProtestDashboard/index.ts
@@ -1,0 +1,3 @@
+import ProtestDashboard from './ProtestDashboard';
+
+export default ProtestDashboard;
